requirements.txt
requests~=2.32.3
beautifulsoup4~=4.13.3
matplotlib~=3.10.1
umap-learn~=0.5.7
streamlit
numpy~=2.1.3
sentence-transformers~=3.4.1
scikit-learn~=1.6.1
torch~=2.6.0